ULTRA-UPDATE

+++ backend-server/application/src/main/java/com/apitable/organization/vo/RoleInfoVo.java
@@ -48,7 +48,7 @@ public class RoleInfoVo {
 
     @Schema(description = "role's member amount", example = "1")
     @JsonSerialize(nullsUsing = NullNumberSerializer.class)
-    private Long memberCount;
+    private Integer memberCount;
 
     @Schema(description = "the role's position in role list", example = "1")
     private Integer position;
