+++ backend-server/application/src/main/java/com/apitable/shared/util/page/PageInfo.java
@@ -18,14 +18,15 @@
 
 package com.apitable.shared.util.page;
 
-import cn.hutool.core.collection.CollUtil;
 import java.util.Collections;
 import java.util.List;
+
+import cn.hutool.core.collection.CollUtil;
 import lombok.Data;
 
 /**
  * <p>
- * page info.
+ * page page
  * </p>
  *
  * @author Shawn Deng
@@ -33,23 +34,23 @@ import lombok.Data;
 @Data
 public class PageInfo<T> {
 
-    private long pageNum;
+    private int pageNum;
 
-    private long pageSize;
+    private int pageSize;
 
-    private long size;
+    private int size;
 
-    private long total;
+    private int total;
 
-    private long pages;
+    private int pages;
 
-    private long startRow;
+    private int startRow;
 
-    private long endRow;
+    private int endRow;
 
-    private long prePage;
+    private int prePage;
 
-    private long nextPage;
+    private int nextPage;
 
     private Boolean firstPage = false;
 
@@ -65,15 +66,7 @@ public class PageInfo<T> {
         //Do Nothing
     }
 
-    /**
-     * constructor.
-     *
-     * @param pageNum  current page number
-     * @param pageSize current page size
-     * @param total    total records
-     * @param records  records
-     */
-    public PageInfo(long pageNum, long pageSize, long total, List<T> records) {
+    public PageInfo(int pageNum, int pageSize, int total, List<T> records) {
         this.pageNum = pageNum;
         this.pageSize = pageSize;
         this.total = total;
@@ -91,7 +84,7 @@ public class PageInfo<T> {
             this.size = 0;
             return;
         }
-        long sub = this.total - ((this.pageNum - 1) * this.pageSize);
+        int sub = this.total - ((this.pageNum - 1) * this.pageSize);
         if (sub <= 0) {
             this.size = 0;
             return;
@@ -101,7 +94,7 @@ public class PageInfo<T> {
             this.size = records.size();
             return;
         }
-        this.records = records.subList(0, (int) sub);
+        this.records = records.subList(0, sub);
         this.size = sub;
     }
 
@@ -109,7 +102,7 @@ public class PageInfo<T> {
         if (getPageSize() == 0) {
             this.pages = 0;
         }
-        long pages = getTotal() / getPageSize();
+        int pages = getTotal() / getPageSize();
         if (getTotal() % getPageSize() != 0) {
             pages++;
         }
@@ -119,7 +112,8 @@ public class PageInfo<T> {
     private void calculateStartAndEndRow() {
         if (this.pageNum > 0) {
             this.startRow = ((this.pageNum - 1) * this.pageSize) + 1;
-        } else {
+        }
+        else {
             this.startRow = 0;
         }
 
@@ -129,12 +123,14 @@ public class PageInfo<T> {
     private void calcPage() {
         if (pageNum > 1) {
             prePage = pageNum - 1;
-        } else {
+        }
+        else {
             prePage = 1;
         }
         if (pageNum < pages) {
             nextPage = pageNum + 1;
-        } else {
+        }
+        else {
             nextPage = pages;
         }
     }
